# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SAMI native runtime",
    "version": "0.2.0",
    "description": "Bounded resource API. See api.md, operation fixtures and source for per-kind validation; generic Record fields are not an exhaustive domain schema."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Local development; use trusted HTTPS in production"
    }
  ],
  "paths": {
...
    "/v1/receipts/{id}/replay": {
      "post": {
        "operationId": "post_v1_receipts_id_replay",
        "summary": "POST /v1/receipts/{id}/replay",
        "description": "Required authority: receipt.read + receipt.replay. Current tenant, purpose, owner and source ACL checks also apply.",
        "responses": {
          "200": {
            "description": "Successful JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or unsupported request"
          },
          "401": {
            "description": "Missing credential"
          },
          "403": {
            "description": "Authority unavailable or denied"
          },
          "404": {
            "description": "Unavailable in this scope"
          },
          "409": {
            "description": "Changed revision, effect binding or state"
          },
          "429": {
            "description": "Rate limit"
          },
          "503": {
            "description": "Required authority or compute unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/v1/export": {
      "get": {
        "operationId": "get_v1_export",
        "summary": "GET /v1/export",
        "description": "Required authority: data.export. Current tenant, purpose, owner and source ACL checks also apply.",
        "responses": {
          "200": {
            "description": "Successful JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or unsupported request"
          },
          "401": {
            "description": "Missing credential"
          },
          "403": {
            "description": "Authority unavailable or denied"
          },
          "404": {
            "description": "Unavailable in this scope"
          },
          "409": {
            "description": "Changed revision, effect binding or state"
          },
          "429": {
            "description": "Rate limit"
          },
          "503": {
            "description": "Required authority or compute unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}